[Kernel/Intrreupts] Added Div 0

+++ kernel/src/Math.h
@@ -4,6 +4,6 @@
 
 struct POINT
 {
-	unsigned int X;
-	unsigned int Y;
+	long X;
+	long Y;
 };
